hw14

+++ src/samples/shadowmap/draw.cpp
@@ -34,7 +34,7 @@ void SimpleShadowmapRender::DrawFrameSimple(bool draw_gui)
   submitInfo.waitSemaphoreCount = 1;
   submitInfo.pWaitSemaphores = waitSemaphores;
   submitInfo.pWaitDstStageMask = waitStages;
-  submitInfo.commandBufferCount = submitCmdBufs.size();
+  submitInfo.commandBufferCount = static_cast<uint32_t>(submitCmdBufs.size());
   submitInfo.pCommandBuffers = submitCmdBufs.data();
 
   VkSemaphore signalSemaphores[] = {m_presentationResources.renderingFinished};
